db config

+++ app.py
@@ -16,7 +16,7 @@ class Expense(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     date = db.Column(db.String(50), nullable=False)
     expensename = db.Column(db.String(50), nullable=False)
-    amount = db.Column(db.Integer(50), nullable=False)
+    amount = db.Column(db.Integer, nullable=False)
     category = db.Column(db.String(50), nullable=False)
